finders/readers: split and clean
This is a first step towards cleaning the finder/reader code. The
end goal is to also clean storage.py and render/views.py to remove
implementation details and make 'remote' a real finder/reader plugin.

- split the various finders/readers but keep importing symbols to
  preserve compatibility.
- run autopep8 on every modified files.
