+++ webapp/graphite/finders/ceres.py
@@ -12,26 +12,28 @@ from . import get_real_metric_path, extract_variants
 
 
 class CeresFinder:
-  def __init__(self, directory=None):
-    directory = directory or settings.CERES_DIR
-    self.directory = directory
-    self.tree = CeresTree(directory)
+    def __init__(self, directory=None):
+        directory = directory or settings.CERES_DIR
+        self.directory = directory
+        self.tree = CeresTree(directory)
 
-  def find_nodes(self, query):
+    def find_nodes(self, query):
 
-    variants = extract_variants(query.pattern)
+        variants = extract_variants(query.pattern)
 
-    for variant in variants:
-      for fs_path in glob( self.tree.getFilesystemPath(variant)):
-        metric_path = self.tree.getNodePath(fs_path)
+        for variant in variants:
+            for fs_path in glob(self.tree.getFilesystemPath(variant)):
+                metric_path = self.tree.getNodePath(fs_path)
 
-        if CeresNode.isNodeDir(fs_path):
-          ceres_node = self.tree.getNode(metric_path)
+                if CeresNode.isNodeDir(fs_path):
+                    ceres_node = self.tree.getNode(metric_path)
 
-          if ceres_node.hasDataForInterval(query.startTime, query.endTime):
-            real_metric_path = get_real_metric_path(fs_path, metric_path)
-            reader = CeresReader(ceres_node, real_metric_path)
-            yield LeafNode(metric_path, reader)
+                    if ceres_node.hasDataForInterval(
+                            query.startTime, query.endTime):
+                        real_metric_path = get_real_metric_path(
+                            fs_path, metric_path)
+                        reader = CeresReader(ceres_node, real_metric_path)
+                        yield LeafNode(metric_path, reader)
 
-        elif os.path.isdir(fs_path):
-          yield BranchNode(metric_path)
+                elif os.path.isdir(fs_path):
+                    yield BranchNode(metric_path)
