+++ webapp/graphite/finders/utils.py
@@ -1,0 +1,12 @@
+"""Utility functions for finders."""
+
+from django.conf import settings
+
+
+def extractForwardHeaders(request):
+    headers = {}
+    for name in settings.REMOTE_STORE_FORWARD_HEADERS:
+        value = request.META.get('HTTP_%s' % name.upper().replace('-', '_'))
+        if value is not None:
+            headers[name] = value
+    return headers
