+++ webapp/graphite/http_pool.py
@@ -1,0 +1,4 @@
+"""Shared urllib3 pool."""
+import urllib3
+
+http = urllib3.PoolManager(num_pools=10, maxsize=5)
