+++ webapp/graphite/readers/__init__.py
@@ -1,0 +1,5 @@
+# Import some symbols to avoid breaking compatibility.
+from graphite.readers.utils import CarbonLink, merge_with_cache, FetchInProgress, MultiReader  # noqa # pylint: disable=unused-import
+from graphite.readers.whisper import WhisperReader, GzippedWhisperReader  # noqa # pylint: disable=unused-import
+from graphite.readers.ceres import CeresReader  # noqa # pylint: disable=unused-import
+from graphite.readers.rrd import RRDReader  # noqa # pylint: disable=unused-import
